Refactor run measurement
